refactor(git): move merge-base tests to mergebase_test.go
